QAM 총합 오류 수정

+++ src/main/java/com/project/erp/fm/model/dto/TransDTO.java
@@ -18,7 +18,7 @@ public class TransDTO {
 	private LocalDate startDate;
 	private LocalDate endDate;
 	private String transMonth;
-	private long monthIncome;
-	private long monthExpenses;
+	private int monthIncome;
+	private int monthExpenses;
 	
 }
